src/PythonApplication1/requirements.txt
pip
jieba
pymilvus
milvus-model
nltk
PyMuPDF
azure-monitor-opentelemetry